Moved tests into main dir, to get rid of relative import madness.
