+++ slider_demo.py
