+++ test_camera.py
